chore: Refactor quality factor level in results.
